% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.info.R
\name{calc.info}
\alias{calc.info}
\title{Compute item information}
\usage{
calc.info(bank, theta)
}
\arguments{
\item{bank}{matrix with item parameters (a, b, c)}

\item{theta}{current theta}
}
\value{
A vector with the information of each item
}
\description{
Calculate information of each item in the bank for a theta
}
\details{

}
\author{
Alexandre Jaloto
}
