% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.prob.R
\name{calc.prob}
\alias{calc.prob}
\title{Compute probability}
\usage{
calc.prob(theta, bank, u = 1)
}
\arguments{
\item{theta}{theta}

\item{bank}{matrix with item parameters (a, b, c)}

\item{u}{\code{1} for correct, \code{0} for wrong}
}
\value{
A vector with the probability of seeing determined response in each item
}
\description{
Calculate probability of observing certain answer to a dichotomous item, given a theta
}
\author{
Alexandre Jaloto
}
