% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.evaluation.R
\name{cat.evaluation}
\alias{cat.evaluation}
\title{CAT Evaluation}
\usage{
cat.evaluation(results, true.scores, item.name, rmax)
}
\arguments{
\item{results}{list with restults of a CAT simulation from \code{simCAT}}

\item{true.scores}{true scores}

\item{item.name}{vector with the name of all items in the bank}

\item{rmax}{item maximum exposure rate}
}
\value{
a list with two elements.

\code{evaluate} is a \code{data.frame}
with the mean of the following variables (in reference
to all replications):
\itemize{
\item \code{rmse} root mean square error between true and estimated score
\item \code{se} standard error of measurement
\item \code{correlation} correlation between true and estimated score
\item \code{bias} bias between true and estimated score
\item \code{overlap} overlap rate
\item \code{min_exp} minimum exposure rate
\item \code{max_exp} maximum exposure rate
\item \code{n_exp0} number of items not administered
\item \code{n_exp_rmax} number of items with exposure rate higher than rmax
\item \code{length_mean} average mean of test length
\item \code{length_sd} standard deviation of test length
\item \code{length_median} average median of test length
\item \code{min_length} minimum test length
\item \code{max_length} maximum test length
}
\code{conditional} is a data.frame with the same variables
conditioned to the true scores. The \code{colnames} are the thetas
in each decil, that is,
\code{quantile(true.scores, probs = seq(.1, 1, length.out = 10))}.
}
\description{
Evaluate a CAT simulation
}
\details{

}
\author{
Alexandre Jaloto
}
