% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eap.R
\name{eap}
\alias{eap}
\title{EAP estimation}
\usage{
eap(pattern, bank)
}
\arguments{
\item{pattern}{response pattern (0 and 1) with the number of columns
corresponding to the number of items}

\item{bank}{data.frame with item parameters (a, b, c)}
}
\value{
data.frame with estimated \code{theta} and \code{SE}.
}
\description{
Estimates theta with Expected a Posteriori
}
\details{
40 quadrature points, ranging from -4 to 4. Priori with normal distribution (mean = 0, sd = 1).
}
\author{
Alexandre Jaloto
}
