% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.rate.R
\name{exposure.rate}
\alias{exposure.rate}
\title{Compute exposure rates}
\usage{
exposure.rate(previous, item.name)
}
\arguments{
\item{previous}{list with previous responses. Each element corresponds to a person and has the names of the applied items.}

\item{item.name}{vector with the name of all items in the bank}
}
\value{
\code{data.frame} with
\itemize{
\item \code{items} name of the items
\item \code{Freq} exposure rate
}
}
\description{
Calculate exposure rate of items in a bank
}
\author{
Alexandre Jaloto
}
