% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.resp.R
\name{gen.resp}
\alias{gen.resp}
\title{Generate response pattern}
\usage{
gen.resp(theta, bank)
}
\arguments{
\item{theta}{theta}

\item{bank}{matrix with item parameters (a, b, c)}
}
\value{
A vector with the probability of seeing determined response in each item
}
\description{
Generate response pattern based on probability of answering correct a dichotomous item, given a theta and an item bank
}
\author{
Alexandre Jaloto
}
