% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Root Mean square Error}
\usage{
rmse(true, estimated)
}
\arguments{
\item{true}{true values}

\item{estimated}{estimated values}
}
\value{
A numeric vector
}
\description{
Calculate the root mean square error
}
\author{
Alexandre Jaloto
}
