% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.item.R
\name{select.item}
\alias{select.item}
\title{Select next item}
\usage{
select.item(
  bank,
  theta,
  administered = NULL,
  sel.method = "MFI",
  stop.type = "precision",
  threshold = 0.3,
  SE,
  acceleration = 1,
  max.items = 45,
  content.names = NULL,
  content.props = NULL,
  content.items = NULL,
  met.content = "MCCAT"
)
}
\arguments{
\item{bank}{matrix with item parameters (a, b, c)}

\item{theta}{current theta}

\item{administered}{vector with administered items, \code{NULL} if it is the first
item}

\item{sel.method}{selection method}

\item{stop.type}{type of stopping rule (\code{precision} or \code{length}).
Necessary only for progressive method.}

\item{threshold}{threshold for \code{stop.type}
Necessary only for progressive method.}

\item{SE}{current standard error
Necessary only for progressive method, with \code{stop.type = "precision"}}

\item{acceleration}{acceleration parameter
#' Necessary only for progressive method.}

\item{max.items}{maximum number of items to be administered
Necessary only for progressive method, with \code{stop.type = "precision"}}

\item{content.names}{vector with the contents of the test}

\item{content.props}{desirable proportion of each content in test, in
the same order of \code{content.names}}

\item{content.items}{vector indicating the content of each item}

\item{met.content}{content balancing method: \code{MCCAT} (default) or \code{CCAT}}
}
\value{
A list with two elements

\code{item} the number o the selected item in item bank
\code{name} name of the selected item (row name)
}
\description{
Select next item to be administered
}
\details{

}
\author{
Alexandre Jaloto
}
