% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.shiny.R
\name{sim.shiny}
\alias{sim.shiny}
\title{CAT simulation in Shiny}
\usage{
sim.shiny()
}
\description{
CAT simulation in a Shiny application.
}
\details{
Uses \code{simCAT} function in a more friendly way. For now, this application
only supports simulation with dichotomous items and one replication.
}
\author{
Alexandre Jaloto
}
