% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCAT.R
\name{simCAT}
\alias{simCAT}
\title{CAT simulation}
\usage{
simCAT(
  resps,
  bank,
  start.theta = 0,
  sel.method = "MFI",
  cat.type = "variable",
  acceleration = 1,
  threshold = 0.3,
  rmax = 1,
  content.names = NULL,
  content.props = NULL,
  content.items = NULL,
  met.content = "MCCAT",
  stop = list(se = 0.3, hypo = 0.015, hyper = Inf)
)
}
\arguments{
\item{resps}{a matrix with responses (0 and 1).
The number of columns
corresponds to the number of items}

\item{bank}{matrix with item parameters (a, b, c)}

\item{start.theta}{first theta}

\item{sel.method}{item selection method}

\item{cat.type}{CAT with \code{variable} or \code{fixed} length
Necessary only for progressive method.}

\item{acceleration}{acceleration parameter.
Necessary only for progressive method.}

\item{threshold}{threshold for \code{cat.type}.
Necessary only for progressive method.}

\item{rmax}{item maximum exposure rate}

\item{content.names}{vector with the contents of the test}

\item{content.props}{desirable proportion of each content in test, in
the same order of \code{content.names}}

\item{content.items}{vector indicating the content of each item}

\item{met.content}{content balancing method: \code{MCCAT} (default) or \code{CCAT}}

\item{stop}{list with stopping rule and thresholds
\itemize{
\item \code{se} minimum standard error
\item \code{delta.theta} minimum absolute difference between current and previous theta
\item \code{hypo} minimum standard error reduction
\item \code{hyper} minimum standard error reduction after achieving \code{se}
\item \code{info} maximum information of an available item
\item \code{max.items} maximum number of items
\item \code{min.items} maximum number of items
\item \code{fixed} fixed number of items
}}
}
\value{

}
\description{
A CAT simulation with dicotomic items.
}
\details{

}
\author{
Alexandre Jaloto
}
